requirements.txt
discord.py>=2.4.0,<3.0.0
aiohttp>=3.9.0,<4.0.0
python-dotenv>=1.0.0,<2.0.0
Pillow>=10.0.0,<12.0.0
tzdata>=2024.1
